new annonce UPDATE

+++ express-project/controllers/Annonce.js
@@ -3,51 +3,69 @@ const Annonce = require('../models/Annonce')
 
  const createAnnonce = async (req, res) => {
     try {
-        const annonce = new Annonce({
-            ...req.body,
-            //user : req.user.id    // recuperer l'id de l'utilisateur qui post l'annonce
-        })
-        await annonce.save()
-        res.status(201).json(annonce)
+        let  annonce  =  req.body ;
+      // Populate the user details
+        const annonceCreated = await  Annonce.create({...annonce ,  user_Id : req.user._id });
+        res.status(201).json(annonceCreated)
     }catch (error){
         res.status(500).json({ message : 'Erreur lors de la creation de l\'annonce' , error : error.message})
     }
 }
-
 const getAll = async (req, res) => {
     try {
         const { status } = req.query;
         const filter = status ? { status } : {};
-        const annonces = await Annonce.find(filter)
+        const annonces = await Annonce.find(filter).populate('user_Id');
         res.status(200).json(annonces)
     } catch (error) {
         res.status(500).json({ message: 'Erreur lors de l\'importation des annonces', error: error.message })
     }
-    /*
-    Get all active annonces: GET /?status=active
-    Get all draft annonces: GET /?status=draft
-    Get all archived annonces: GET /?status=archived 
-    */
 }
-
 const getById = async (req, res) => {
     try{
-        const annonce = await Annonce.findById(req.params.id)
+        const annonce = await Annonce.findById(req.params.id).populate('user_Id');
+        if(!annonce) {
+         return  res.status(404).json({
+            message : "Annonced Id is wrong !"
+          })
+        }
         res.status(200).json(annonce)
     } catch (error) {
         res.status(500).json({ message : 'Erreur lors de l\'importation des annonces' , error : error.message})
     }
 }
-
 const updateAnnonce = async (req, res) => {
     try {
+
+        const oldAnnonce = await Annonce.findById(req.params.id);
+        if( !oldAnnonce ) {
+          return res.status(404).json({
+            message : "Annonce is not valid or has been deleted "
+          })
+        }
+
         const annonce = await Annonce.findByIdAndUpdate(req.params.id, req.body, { new: true });
-        res.status(200).json(annonce);
+        if(!annonce){
+          return res.status(500).json({
+            message :"Erreur lors de mise à jour de  l'annonce",
+          })
+        }
+        res.status(200).json( {
+          message : "new Annonce has been updated " ,
+          newAnnonce : annonce
+        });
     } catch (error) {
         res.status(500).json({ message : 'Erreur lors de la mise à jour de l\'annonce', error : error.message });
     }
-};const deleteAnnonce = async (req, res) => {
+};
+const deleteAnnonce = async (req, res) => {
     try {
+        const annonce = await Annonce.findById(req.params.id);
+        if(!annonce) {
+          return res.status(404).json({
+            message : "Annonce is not valid or has been deleted "
+          })
+        }
         await Annonce.findByIdAndRemove(req.params.id);
         res.status(200).json({ message: 'Annonce supprimée avec succès' });
     } catch (error) {
